backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
qdrant-client==1.16.2
pydantic==2.9.2
google-generativeai==0.8.4
python-dotenv==1.0.1
asyncpg==0.29.0
sqlalchemy==2.0.35
alembic==1.13.3
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
python-slugify==8.0.1